Size-ratchet fit: comment compression only, sweep test in its own module
delegate_custody returns to the 1600 cap through comment/docstring
compression alone; the registration-sweep deferral test moves to its own
module instead of pushing the isolation suite over the cap; the transport
suite returns under its byte pin; agent_startup_checks re-enters the
1001-1500 band with the reduction rationale (the envelope producer moved
to the contracts SSOT).
