+++ ouroboros/delegate_custody.py
@@ -61,13 +61,12 @@ CONTAINMENT_FAULT = "delegate_run_containment_fault"
 CONTAINMENT_RESOLVED = "delegate_run_containment_resolved"
 RECONCILED = "delegate_run_reconciled"
 OUTPUT_SPILLED = "delegate_run_output_spilled"
-# Owner doctrine D7: a delegated result is OBTAINED only after the artifact is read to
-# EOF. The canonical acknowledgement — written when the read windows covered the staged
-# artifact CONTINUOUSLY start-to-EOF, hash-bound, and only writable when the staging was
-# the verified FULL content (`full_content`), never the bounded 256 KiB preview.
-# Owner directive (2026-08-03): consumption is LOAD-BEARING before settlement; its
-# ABSENCE is the durable typed fact ``SETTLED_UNREAD`` below. It does not BLOCK
-# settlement — see ``settle_run`` for why a hard gate would deadlock the flow.
+# Owner doctrine D7: a result is OBTAINED only after the artifact is read to
+# EOF - the ack is written when read windows covered the staged artifact
+# CONTINUOUSLY start-to-EOF, hash-bound, and only over verified FULL content
+# (`full_content`), never the bounded 256 KiB preview. Owner directive
+# (2026-08-03): consumption is LOAD-BEARING before settlement; its ABSENCE is
+# ``SETTLED_UNREAD`` below, and it does not BLOCK settlement (``settle_run``).
 OUTPUT_CONSUMED = "delegate_run_output_consumed"
 # A run whose VERIFIED FULL output was staged, paid for, and settled without ever being
 # read to EOF: the "launched and never collected" class, named at the moment it becomes
@@ -146,41 +145,38 @@ class RunCustody:
     authority_fingerprint: str = ""
     ledger_recorded: bool = False
     settled: bool = False
-    # The containment disclosure is a FACT about the run, so it is written once. A nanny
-    # may call `delegate_wait` again on an already-terminal run, and a durable log that
-    # repeats the same finding per poll reads like repeated findings. Replayed from the
-    # durable row like every other field here, so a restarted worker does not re-disclose.
+    # The containment disclosure is a FACT about the run, written once (a nanny
+    # re-polling a terminal run must not read as repeated findings); replayed
+    # durably, so a restarted worker does not re-disclose.
     containment_disclosed: bool = False
     # Whether the settled-but-never-read omission has already been named durably.
     unread_disclosed: bool = False
-    # The staged-output half of the terminal story (D7). ``output_artifact`` is the task-drive-relative path the
-    # terminal payload was staged to (empty when it fit inline); ``output_complete`` is whether what was staged is
-    # the full content as far as the run's OWN report can establish it — its served size, or the preview carried
-    # as a prefix (`_resolve_full_primary_output`); the engine publishes no content hash for the primary output,
-    # so equal length binds the body to the run's CLAIM about it, not to a digest of it. An artifact that matches
-    # neither stages as incomplete and can never be acknowledged. ``output_sha`` is the content hash of what is
-    # CURRENTLY staged; ``output_consumed`` is whether the canonical acknowledgement exists FOR THAT HASH — a
-    # re-stage of different content at the same path resets it, because an ack names bytes, not a path. All
-    # replayed, so a restarted worker keeps the facts.
+    # The staged-output half of the terminal story (D7). ``output_artifact``: the
+    # task-drive-relative staging path (empty when inline). ``output_complete``:
+    # whether the staged body matches the run's OWN report - served size or the
+    # carried preview prefix (`_resolve_full_primary_output`); the engine
+    # publishes no content hash, so equal length binds to the run's CLAIM, and
+    # an artifact matching neither stages incomplete and can never be
+    # acknowledged. ``output_sha``: hash of what is CURRENTLY staged;
+    # ``output_consumed``: whether the canonical ack exists FOR THAT HASH (an
+    # ack names bytes, not a path). All replayed across worker restarts.
     output_artifact: str = ""
     output_complete: bool = False
     output_sha: str = ""
     output_consumed: bool = False
-    # C1 isolation binding for a MUTATING run: the private execution root the run was
-    # scoped to, the baseline commit its diff is measured from, and the AUTHORITY
-    # target tree its patch is destined for. ``snapshot_id`` keys the worktree-service
-    # registry entry (it equals the invocation id that provisioned it). All durable and
-    # replayed: a retry reproduces the exact binding, the terminal capture knows where
-    # to diff, and the startup GC can tell a live snapshot from a disposable one.
+    # C1 isolation binding for a MUTATING run: private execution root, diff
+    # baseline commit, AUTHORITY target tree. ``snapshot_id`` keys the
+    # worktree-service registry entry (= the provisioning invocation id). All
+    # durable/replayed: retries reproduce the binding, terminal capture knows
+    # where to diff, startup GC tells live snapshots from disposable ones.
     snapshot_id: str = ""
     execution_root: str = ""
     baseline_sha: str = ""
     target_root: str = ""
     authority_source: str = ""
-    # The GRANTED run shape, replayed off the STARTED row that already carries it
-    # (the `shape` dict): delegate_wait replays this as the entitled authority —
-    # re-deriving from live context read `readonly` for a top-level payload run
-    # and cancelled it as widened (R1-2).
+    # The GRANTED run shape, replayed off the STARTED row's `shape` dict:
+    # delegate_wait replays this as the entitled authority - re-deriving from
+    # live context read `readonly` and cancelled the run as widened (R1-2).
     access: str = ""
     mode: str = ""
     isolation: str = ""
@@ -817,16 +813,13 @@ def retire_project(drive_root: Any, gateway: Any, custody: RunCustody) -> None:
     if not (custody.project_owned and custody.project_id):
         return
     try:
-        # Sharers = EVERY run in the project, owned or not (only the creator
+        # Sharers = EVERY run in the project, owned or not: only the creator
         # carries project_owned, but the daemon refuses removal while ANY
-        # sibling lives). The retry lane is keyed on the REGISTRATION, not
-        # settlement: keying on ``not settled`` made settled sharers
-        # invisible - nobody carried the retry lane.
+        # sibling lives. The caller is mid-settlement (its SETTLED emit
+        # follows this call), so only OTHER unsettled sharers defer the
+        # attempt - the daemon would refuse it anyway.
         rows = [run for run in replay(drive_root).values()
                 if run.project_id == custody.project_id and run.run_id]
-        # The caller is mid-settlement (its own SETTLED emit follows this
-        # call), so its replayed row still reads unsettled - only OTHER
-        # unsettled sharers defer the attempt: the daemon would refuse.
         if any(not run.settled and run.run_id != custody.run_id for run in rows):
             return
         sharers = sorted(run.run_id for run in rows if run.project_owned)
@@ -889,11 +882,10 @@ def settle_run(drive_root: Any, gateway: Any, custody: RunCustody, detail: Dict[
         return {"settled": True, "ledger_recorded": True,
                 "project_retired": not custody.project_owned, "retried": False}
     summary = summary_of(detail)
-    # Claudexor reports CASH in `spendUsd` and its EXACTNESS in `spendEstimated`. A
-    # delegated run is only free when the amount is really zero AND really settled: an
-    # expired session, a route that bills by construction, or an auth fallback all
-    # produce a real charge, and writing 0.0/cost_final=True over any of them would hide
-    # the money from every budget fence AND assert the projection is final.
+    # Claudexor reports CASH in `spendUsd`, EXACTNESS in `spendEstimated`. A run
+    # is only free when the amount is really zero AND really settled: expired
+    # sessions, bill-by-construction routes and auth fallbacks all charge, and
+    # writing 0.0/cost_final=True over them hides money from every budget fence.
     spend, estimated = disclosed_spend(summary)
     # D29: the applied credential-profile id + access profile the deciding
     # attempt disclosed (authRoute receipt / effectiveAccess), written to the
@@ -966,15 +958,12 @@ def settle_run(drive_root: Any, gateway: Any, custody: RunCustody, detail: Dict[
         })
     if custody.settled:
         resolve_containment_fault(drive_root, custody, "settled_terminal")
-    # CONSUMPTION BEFORE SETTLEMENT is the owner's directive, and it is enforced as a
-    # LOUD FACT rather than a gate — but NOT from here. This function runs BEFORE the
-    # artifact is staged (``delegate_wait`` settles, then builds the payload that stages
-    # the file), so at this point there is nothing staged to be unread yet: asking the
-    # question here would answer "no omission" for every first settlement, which is the
-    # render-unknown-as-a-fact shape this module refuses everywhere else. The fact is
-    # recorded by ``record_settled_unread`` at the two places that DO know it — the wait
-    # path once staging is done, and reconciliation, whose custody replays with the
-    # staged fields already on it.
+    # CONSUMPTION BEFORE SETTLEMENT is the owner's directive - a LOUD FACT, not
+    # a gate, and NOT recorded here: this runs BEFORE staging (delegate_wait
+    # settles, then stages), so asking now would answer "no omission" for every
+    # first settlement - the render-unknown-as-a-fact shape this module refuses.
+    # ``record_settled_unread`` records it at the two places that DO know: the
+    # wait path after staging, and reconciliation (custody replays staged).
     return {
         "settled": custody.settled,
         "ledger_recorded": custody.ledger_recorded,
@@ -1237,10 +1226,9 @@ def owned_project_registrations(drive_root: Any) -> List[RunCustody]:
 
 
 def retire_settled_registrations(drive_root: Any, gateway: Any) -> None:
-    """Retire projects every sharer has settled; a LIVE sharer defers the
-    attempt (the daemon would refuse anyway). Sharers are ALL runs in the
-    project, owned or not - only the creator carries the registration, but
-    any live sibling makes the daemon refuse. Idempotent, fail-soft."""
+    """Retire projects every sharer has settled; a LIVE sharer (owned or not
+    - only the creator carries the registration, but any live sibling makes
+    the daemon refuse) defers the attempt. Idempotent, fail-soft."""
     by_project: Dict[str, List[RunCustody]] = {}
     for row in replay(drive_root).values():
         if row.project_id and row.run_id:
@@ -1248,7 +1236,7 @@ def retire_settled_registrations(drive_root: Any, gateway: Any) -> None:
     for rows in by_project.values():
         owned = [row for row in rows if row.project_owned]
         if not owned or any(not row.settled for row in rows):
-            continue
+            continue  # nothing registered here, or a live sharer defers
         try:
             retire_project(drive_root, gateway, min(owned, key=lambda row: row.run_id))
         except Exception:
@@ -1346,9 +1334,9 @@ def _reconcile_each(drive_root: Any, runs: List[RunCustody],
     startup/periodic sweep surface is where settled-but-owned registrations get
     their retry lane (the task-scoped surface can exit early with none open).
     """
-    # Registration duty counts as real work only when a project is actually
-    # retire-ELIGIBLE (every sharer settled): a registration deferred behind a
-    # live sibling must not spin the owned daemon up on every sweep tick.
+    # Registration duty is real work only when a project is retire-ELIGIBLE
+    # (every sharer settled): a registration deferred behind a live sibling
+    # must not spin the owned daemon up on every sweep tick.
     snapshot = replay(drive_root).values()
     unsettled_projects = {row.project_id for row in snapshot
                           if row.project_id and row.run_id and not row.settled}
@@ -1360,12 +1348,11 @@ def _reconcile_each(drive_root: Any, runs: List[RunCustody],
     from ouroboros.gateways.claudexor import ClaudexorUnavailable
 
     if gateway_factory is None:
-        # The startup sweep REAPS the previous generation's owned daemon right before
-        # calling here, so a bare discovery-only gateway always found a corpse and the
-        # whole reconciliation silently no-opped on every restart — open runs stayed
-        # unsettled until the next delegate_start happened to revive the daemon. The
-        # ensure path starts our own daemon when there is real work to reconcile
-        # (never on the empty early-return above), and as a side effect activates a
+        # The startup sweep REAPS the previous generation's owned daemon just
+        # before calling here, so a bare discovery-only gateway always found a
+        # corpse and reconciliation silently no-opped on every restart. The
+        # ensure path starts our own daemon when there is real work to
+        # reconcile (never on the empty early-return above), activating a
         # staged runtime update the old always-running daemon could never adopt.
         from ouroboros.claudexor_daemon import ensure_owned_gateway
 
@@ -1382,8 +1369,8 @@ def _reconcile_each(drive_root: Any, runs: List[RunCustody],
             outcomes.append(_reconcile_one(drive_root, gateway, custody))
         for record in pending or []:
             outcomes.append(_recover_pending_invocation(drive_root, gateway, record))
-        # Recomputed inside: a run this very pass just settled may have made
-        # its project eligible, so the pre-pass gate is not the last word.
+        # Recomputed inside: a run settled this very pass may have made its
+        # project eligible - the pre-pass gate is not the last word.
         if registrations or runs:
             retire_settled_registrations(drive_root, gateway)
     finally:
@@ -1460,11 +1447,10 @@ def _recover_pending_invocation(drive_root: Any, gateway: Any,
             if isinstance(record.get("work_order_source_request"), dict) else {}
         ),
         authority_fingerprint=str(record.get("authority_fingerprint") or ""),
-        # The C1 isolation binding survives recovery VERBATIM: the recovered run
-        # executes in the snapshot the original attempt provisioned (the replayed
+        # The C1 isolation binding survives recovery VERBATIM: the recovered
+        # run executes in the originally provisioned snapshot (the replayed
         # body's scope.root), so its STARTED row must name that binding or the
-        # snapshot — and the child's work in it — becomes GC food the moment the
-        # invocation stops being pending.
+        # snapshot and the child's work become GC food once no longer pending.
         snapshot_id=str(record.get("snapshot_id") or ""),
         execution_root=str(record.get("execution_root") or ""),
         baseline_sha=str(record.get("baseline_sha") or ""),
@@ -1517,14 +1503,12 @@ def _reconcile_one(drive_root: Any, gateway: Any, custody: RunCustody) -> Dict[s
         else:
             record_containment_fault(drive_root, custody, "reconcile_unreadable", f"{exc.code}: {exc}")
             result = {"run_id": custody.run_id, "task_id": custody.task_id, "action": "unreadable"}
-        # NO capture here (C1-R2): an absent run's state is unknowable from this
-        # daemon — across the D30 owned-daemon provisioning boundary the child may
-        # still be alive and WRITING to the snapshot, and an eager capture would
-        # freeze a potentially incomplete patch which the idempotent capture core
-        # would then serve forever. Custody closes, the snapshot stays preserved
-        # (undisposed, so the GC keeps it), the obligation surfaces through
-        # ``undisposed_patches()``, and the capture happens at disposition
-        # (``integrate_delegated_patch``) — the honest latest-possible point.
+        # NO capture here (C1-R2): across the D30 provisioning boundary an
+        # absent run may still be WRITING its snapshot, and an eager capture
+        # would freeze an incomplete patch the idempotent core then serves
+        # forever. Custody closes, the snapshot survives (undisposed - GC keeps
+        # it), the duty surfaces via ``undisposed_patches()``, and capture
+        # happens at disposition (``integrate_delegated_patch``).
         emit(drive_root, RECONCILED, result)
         return result
     if is_terminal(detail):
